:bug: When open auth renew priority judgment

+++ bot-controller/src/main/java/com/ocd/controller/task/TimerTask.java
@@ -80,31 +80,32 @@ public class TimerTask {
             Long betweenPlayDay = activityLogs.isEmpty() ? null : DateUtil.betweenDay(activityLogs.get(0).getDateCreated(), new Date(), true);
             Long betweenExpDay = DateUtil.betweenDay(user.getExpTime(), new Date(), true);
             String lastDate = activityLogs.isEmpty() ? "无" : FormatUtil.INSTANCE.dateToString(activityLogs.get(0).getDateCreated());
-            if (AuthorityUtil.botConfig.getCleanTask() && (betweenPlayDay == null || betweenPlayDay >= AuthorityUtil.botConfig.getExpDay())) {
-                if (AuthorityUtil.botConfig.getDelete() && (betweenPlayDay == null || betweenPlayDay >= AuthorityUtil.botConfig.getExpDelDay() + AuthorityUtil.botConfig.getExpDelDay())) {
-                    EmbyUtil.getInstance().deleteUser(user);
-                    user.cleanEmby();
-                } else {
-                    EmbyUtil.getInstance().deactivateUser(user, true);
-                    user.setDeactivate(true);
-                }
-                sendMessage.enableMarkdownV2(true);
-                sendMessage.setChatId(AuthorityUtil.botConfig.notifyChannel);
-                sendMessage.setText(MessageUtil.INSTANCE.getAccountMessage(user, lastDate, true));
-                try {
-                    telegramClient.execute(sendMessage);
-                } catch (TelegramApiException e) {
-                    log.error(e.toString());
-                } finally {
-                    sendMessage.setChatId(user.getTgId());
+            if (AuthorityUtil.botConfig.getOpenAutoRenewal() && expDate.before(user.getExpTime()) || !AuthorityUtil.botConfig.getOpenAutoRenewal())
+                if (AuthorityUtil.botConfig.getCleanTask() && (betweenPlayDay == null || betweenPlayDay >= AuthorityUtil.botConfig.getExpDay())) {
+                    if (AuthorityUtil.botConfig.getDelete() && (betweenPlayDay == null || betweenPlayDay >= AuthorityUtil.botConfig.getExpDelDay() + AuthorityUtil.botConfig.getExpDelDay())) {
+                        EmbyUtil.getInstance().deleteUser(user);
+                        user.cleanEmby();
+                    } else {
+                        EmbyUtil.getInstance().deactivateUser(user, true);
+                        user.setDeactivate(true);
+                    }
+                    sendMessage.enableMarkdownV2(true);
+                    sendMessage.setChatId(AuthorityUtil.botConfig.notifyChannel);
+                    sendMessage.setText(MessageUtil.INSTANCE.getAccountMessage(user, lastDate, true));
                     try {
                         telegramClient.execute(sendMessage);
                     } catch (TelegramApiException e) {
                         log.error(e.toString());
+                    } finally {
+                        sendMessage.setChatId(user.getTgId());
+                        try {
+                            telegramClient.execute(sendMessage);
+                        } catch (TelegramApiException e) {
+                            log.error(e.toString());
+                        }
                     }
+                    return;
                 }
-                return;
-            }
             if (AuthorityUtil.botConfig.getOpenAutoRenewal() && expDate.before(user.getExpTime())) {
                 if (user.getPoints() < AuthorityUtil.botConfig.getUnblockPoints()) {
                     if (AuthorityUtil.botConfig.getDelete() && betweenExpDay >= AuthorityUtil.botConfig.getExpDelDay()) {
